Refactor test configs into separate classes
Moved DefaultConfig and FullConfig inner classes from RedisReplicatorConfigurationTest to new top-level classes: DefaultRedisReplicationConfig and FullRedisReplicationConfig. Updated test references to use the new config classes for improved clarity and maintainability.
