+++ microsphere-redis-replicator-spring/src/test/java/io/microsphere/redis/replicator/spring/config/FullRedisReplicationConfig.java
@@ -1,0 +1,78 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.microsphere.redis.replicator.spring.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.StringRedisTemplate;
+
+/**
+ * Full Redis Replication Config
+ *
+ * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
+ * @see DefaultRedisReplicationConfig
+ * @since 1.0.0
+ */
+@PropertySource(
+        RedisReplicatorConfigurationTest.TEST_PROPERTY_SOURCE_LOCATION
+)
+public class FullRedisReplicationConfig extends DefaultRedisReplicationConfig {
+
+    @Bean
+    public RedisTemplate defaultRedisTemplate(RedisConnectionFactory redisConnectionFactory) {
+        return newRedisTemplate(redisConnectionFactory);
+    }
+
+    @Bean
+    public StringRedisTemplate defaultStringRedisTemplate(RedisConnectionFactory redisConnectionFactory) {
+        return newStringRedisTemplate(redisConnectionFactory);
+    }
+
+    @Bean
+    public RedisTemplate testRedisTemplate(RedisConnectionFactory redisConnectionFactory) {
+        return newRedisTemplate(redisConnectionFactory);
+    }
+
+    @Bean
+    public StringRedisTemplate testStringRedisTemplate(RedisConnectionFactory redisConnectionFactory) {
+        return newStringRedisTemplate(redisConnectionFactory);
+    }
+
+    RedisTemplate newRedisTemplate(RedisConnectionFactory redisConnectionFactory) {
+        RedisTemplate redisTemplate = new RedisTemplate();
+        redisTemplate.setConnectionFactory(redisConnectionFactory);
+        return redisTemplate;
+    }
+
+    StringRedisTemplate newStringRedisTemplate(RedisConnectionFactory redisConnectionFactory) {
+        StringRedisTemplate stringRedisTemplate = new StringRedisTemplate();
+        stringRedisTemplate.setConnectionFactory(redisConnectionFactory);
+        return stringRedisTemplate;
+    }
+
+    @Bean
+    public RedisConnectionFactory redisConnectionFactory() {
+        LettuceConnectionFactory redisConnectionFactory = new LettuceConnectionFactory("127.0.0.1", 6379);
+        redisConnectionFactory.afterPropertiesSet();
+        redisConnectionFactory.validateConnection();
+        return redisConnectionFactory;
+    }
+}
